Changed UFFS_ERR_xxx to UFFS_MSG_xxx; Fixed compile problems on very old compiler

+++ src/inc/uffs/uffs_blockinfo.h
@@ -51,7 +51,7 @@ extern "C"{
  */
 struct uffs_PageSpareSt {
 	uffs_Tags tag;			//!< page tag
-	u8 expired:1;
+	int expired:1;
 };
 
 /**
